Move testUtil.js to simplify ava command line.
