+++ ohm-js/package.json
@@ -29,8 +29,8 @@
     "clean": "rm -f dist/ohm.js dist/ohm.min.js",
     "lint": "eslint . --ignore-path ../.eslintignore",
     "pretest": "bash ../bin/update-env.sh",
-    "test": "ava 'test/**/test-*.js' && ava --config ava-ts.config.js test/test-typings.ts",
-    "test-watch": "ava 'test/**/test-*.js' --watch",
+    "test": "ava && ava --config ava-ts.config.js test/test-typings.ts",
+    "test-watch": "ava --watch",
     "pre-commit": "yarn run lint && yarn run build && yarn run test",
     "prepublishOnly": "cp ../README.md .",
     "postpublish": "rm README.md",
