+++ ohm-js/test/helpers/testUtil.js
@@ -4,7 +4,7 @@
 // Imports
 // --------------------------------------------------------------------
 
-const ohm = require('..');
+const ohm = require('../..');
 
 // --------------------------------------------------------------------
 // Private stuff
