+++ ohm-js/test/test-incremental.js
@@ -7,7 +7,7 @@
 const fs = require('fs');
 const ohm = require('..');
 const test = require('ava');
-const testUtil = require('./testUtil');
+const testUtil = require('./helpers/testUtil');
 
 const makeGrammar = testUtil.makeGrammar;
