+++ ohm-js/test/test-pexprs-generateExample.js
@@ -2,7 +2,7 @@
 
 const test = require('ava');
 
-const makeGrammar = require('./testUtil').makeGrammar;
+const makeGrammar = require('./helpers/testUtil').makeGrammar;
 
 // --------------------------------------------------------------------
 // Tests
