+++ ohm-js/test/test-semantics.js
@@ -8,7 +8,7 @@ const fs = require('fs');
 const test = require('ava');
 
 const ohm = require('..');
-const testUtil = require('./testUtil');
+const testUtil = require('./helpers/testUtil');
 
 // --------------------------------------------------------------------
 // Helpers
